requirements.txt
Django==4.2.5
langchain==0.0.329
google-generativeai==0.3.1
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
PyAudio==0.2.11
